wip meters updates
