+++ frontend/src/components/ui/index.ts
@@ -22,6 +22,7 @@ import("./details");
 import("./file-input");
 import("./file-list");
 import("./filter-chip");
+import("./floating-popover");
 import("./format-date");
 import("./inline-input");
 import("./language-select");
