+++ frontend/src/features/meters/colors.ts
@@ -1,0 +1,11 @@
+import { tw } from "@/utils/tailwind";
+
+export const colors = {
+  default: tw`text-neutral-600`,
+  crawls: tw`text-lime-500`,
+  uploads: tw`text-sky-500`,
+  archivedItems: tw`text-primary-500`,
+  browserProfiles: tw`text-orange-500`,
+  runningTime: tw`text-blue-600`,
+  misc: tw`text-gray-400`,
+};
