+++ frontend/src/features/meters/index.ts
@@ -1,0 +1,2 @@
+import "./execution-minute-meter";
+import "./storage-meter";
